Moved re-used code pieces to separate info_card.dart file
