+++ lib/pages/about.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../features/header.dart';
+import '../features/info_card.dart';
 
 class About extends StatelessWidget {
   const About({super.key});
@@ -47,7 +48,23 @@ class About extends StatelessWidget {
                     flex: 5,
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [AboutCard(), AboutCard(), AboutCard()],
+                      children: [
+                        InfoCard(
+                            icon: Icons.info,
+                            title: "About",
+                            description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                        ),
+                        InfoCard(
+                            icon: Icons.info,
+                            title: "About",
+                            description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                        ),
+                        InfoCard(
+                            icon: Icons.info,
+                            title: "About",
+                            description: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+                        ),
+                      ],
                     ),
                   ),
                 ],
@@ -62,38 +79,6 @@ class About extends StatelessWidget {
             child: Text(
               "Factcheck Yourself",
               style: TextStyle(fontSize: 48, fontWeight: FontWeight.bold),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class AboutCard extends StatelessWidget {
-  const AboutCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.all(16),
-      child: Row(
-        children: [
-          Icon(Icons.info, color: Colors.grey[600]),
-          const SizedBox(width: 16),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Text(
-                  "About",
-                  style: TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-                ),
-                const SizedBox(height: 16),
-                Text(
-                  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-                ),
-              ],
             ),
           ),
         ],
